+++ jemoc_stream/src/main/cpp/include/stream/MemfdStream.h
@@ -23,6 +23,7 @@ public:
     void flush() override;
     void close() override;
     void setLength(long length) override;
+    void sendFile(const int& fd);
 
     // 获取 memfd 的文件描述符
     int getFd() const;
@@ -35,6 +36,7 @@ public:
     static napi_value JSToArrayBuffer(napi_env env, napi_callback_info info);
     static void Export(napi_env env, napi_value exports);
     static napi_value JSGetFd(napi_env env, napi_callback_info info);
+    static napi_value JSSendFile(napi_env env, napi_callback_info info);
     napi_value readAllFromFd(napi_env env);
 
 private: